feat(search): replace legacy SearchHeader by SearchHeaderRework
